Change 1 and suggestion
eof() is set after extraction of data from the file. You cannot use it like that in a loop. the first time eof() is called in the while loop it is not defined. Refer to this:
http://stackoverflow.com/questions/5605125/why-is-iostreameof-inside-a-loop-condition-considered-wrong

Also, I have found a great source for huge text files which can be very useful for your program:
http://introcs.cs.princeton.edu/java/data/
For example a good data set from this site is the words in the complete works of Shakespeare with 29166 words! This is the URL:
http://introcs.cs.princeton.edu/java/data/words.shakespeare.txt